+++ autosub/__init__.py
@@ -6,284 +6,136 @@ Defines autosub's main functionality.
 
 from __future__ import absolute_import, print_function, unicode_literals
 
+# Import built-in modules
 import argparse
-import audioop
-import math
 import multiprocessing
 import os
-import subprocess
-import tempfile
 import wave
-import re
-import json
-import requests
+
+# Import third-party modules
 import pysubs2
 import auditok
-try:
-    from json.decoder import JSONDecodeError
-except ImportError:
-    JSONDecodeError = ValueError
+import progressbar
 
-from googleapiclient.discovery import build
-from progressbar import ProgressBar, Percentage, Bar, ETA
-
+# Any changes to the path and your own modules
 from autosub import constants
 from autosub import formatters
 from autosub import metadata
+from autosub import speech_trans_api
+from autosub import ffmpeg_utils
 
 
-def percentile(arr, percent):
-    """
-    Calculate the given percentile of arr.
-    """
-    arr = sorted(arr)
-    index = (len(arr) - 1) * percent
-    floor = math.floor(index)
-    ceil = math.ceil(index)
-    if floor == ceil:
-        return arr[int(index)]
-    low_value = arr[int(floor)] * (ceil - index)
-    high_value = arr[int(ceil)] * (index - floor)
-    return low_value + high_value
-
-
-class SplitIntoFLACPiece(object): # pylint: disable=too-few-public-methods
-    """
-    Class for converting a region of an input audio or video file into a FLAC audio file
-    """
-
-    def __init__(self, source_path, include_before=0.25, include_after=0.25):
-        self.source_path = source_path
-        self.include_before = include_before
-        self.include_after = include_after
-
-    def __call__(self, region):
-        try:
-            start_ms, end_ms = region
-            start = float(start_ms) / 1000.0
-            end = float(end_ms) / 1000.0
-            start = max(0.0, start - self.include_before)
-            end += self.include_after
-            temp = tempfile.NamedTemporaryFile(suffix='.flac', delete=False)
-            command = ["ffmpeg", "-ss", str(start), "-t", str(end - start),
-                       "-y", "-i", self.source_path, "-c", "copy",
-                       "-loglevel", "error", temp.name]
-            use_shell = True if os.name == "nt" else False
-            subprocess.check_output(command, stdin=open(os.devnull), shell=use_shell)
-            read_data = temp.read()
-            temp.close()
-            os.unlink(temp.name)
-            return read_data
-
-        except KeyboardInterrupt:
-            return None
-
-
-class GoogleSpeechToTextV2(object): # pylint: disable=too-few-public-methods
-    """
-    Class for performing speech-to-text for an input FLAC file.
-    """
-    def __init__(self, api_url, language="en",
-                 rate=44100, retries=3, api_key=constants.GOOGLE_SPEECH_V2_API_KEY):
-        # pylint: disable=too-many-arguments
-        self.language = language
-        self.rate = rate
-        self.api_url = api_url
-        self.api_key = api_key
-        self.retries = retries
-
-    def __call__(self, data):
-        try:
-            for _ in range(self.retries):
-                url = self.api_url.format(lang=self.language, key=self.api_key)
-                headers = {"Content-Type": "audio/x-flac; rate=%d" % self.rate}
-
-                try:
-                    resp = requests.post(url, data=data, headers=headers)
-                except requests.exceptions.ConnectionError:
-                    continue
-
-                for line in resp.content.decode('utf-8').split("\n"):
-                    try:
-                        line = json.loads(line)
-                        line = line['result'][0]['alternative'][0]['transcript']
-                        return line[:1].upper() + line[1:]
-                    except (JSONDecodeError, ValueError, IndexError):
-                        # no result
-                        continue
-
-        except KeyboardInterrupt:
-            return None
-
-
-class GoogleTranslatorV2(object):  # pylint: disable=too-few-public-methods
-    """
-    Class for translating a sentence from a one language to another.
-    """
-    def __init__(self, language, api_key, src, dst):
-        self.language = language
-        self.api_key = api_key
-        self.service = build('translate', 'v2',
-                             developerKey=self.api_key)
-        self.src = src
-        self.dst = dst
-
-    def __call__(self, sentence):
-        try:
-            if not sentence:
-                return None
-
-            result = self.service.translations().list( # pylint: disable=no-member
-                source=self.src,
-                target=self.dst,
-                q=[sentence]
-            ).execute()
-
-            if 'translations' in result and result['translations'] and \
-                'translatedText' in result['translations'][0]:
-                return result['translations'][0]['translatedText']
-
-            return None
-
-        except KeyboardInterrupt:
-            return None
-
-
-def which_exe(program):
-    """
-    Return the path for a given executable.
-    """
-    def is_exe(file_path):
-        """
-        Checks whether a file is executable.
-        """
-        return os.path.isfile(file_path) and os.access(file_path, os.X_OK)
-
-    fpath, _ = os.path.split(program)
-    if fpath:
-        if is_exe(program):
-            return program
-    else:
-        for path in os.environ["PATH"].split(os.pathsep):
-            path = path.strip('"')
-            exe_file = os.path.join(path, program)
-            if is_exe(exe_file):
-                return exe_file
-    return None
-
-
-def ffmpeg_check():
-    """
-    Return the ffmpeg executable name. "None" returned when no executable exists.
-    """
-    if which_exe("ffmpeg"):
-        return "ffmpeg"
-    if which_exe("ffmpeg.exe"):
-        return "ffmpeg.exe"
-    return None
-
-
-def source_to_audio(filename, channels=1,
-                    rate=48000, file_ext='.wav',
-                    ffmpeg_loglevel='error'):
-    """
-    Convert input file to a temporary audio file.
-    """
-    temp = tempfile.NamedTemporaryFile(suffix=file_ext, delete=False)
-    if not os.path.isfile(filename):
-        print("The given file does not exist: {}.".format(filename))
-        raise Exception("Invalid filepath: {}.".format(filename))
-    if not ffmpeg_check():
-        print("ffmpeg: Executable not found on this machine.")
-        raise Exception("Dependency not found: ffmpeg.")
-    command = [ffmpeg_check(), "-y", "-i", filename,
-               "-ac", str(channels), "-ar", str(rate),
-               "-loglevel", ffmpeg_loglevel, temp.name]
-    use_shell = True if os.name == "nt" else False
-    subprocess.check_output(command, stdin=open(os.devnull), shell=use_shell)
-    return temp.name
-
-
-def generate_subtitles( # pylint: disable=too-many-locals,too-many-arguments,too-many-branches,too-many-statements
+def auditok_gen_speech_regions(  # pylint: disable=too-many-arguments
         source_file,
-        output=None,
-        concurrency=constants.DEFAULT_CONCURRENCY,
-        src_language=constants.DEFAULT_SRC_LANGUAGE,
-        dst_language=constants.DEFAULT_DST_LANGUAGE,
-        subtitles_file_format=constants.DEFAULT_SUBTITLES_FORMAT,
-        api_url_scheme=constants.DEFAULT_API_URL_SCHEME,
-        api_key=None,
+        energy_threshold=constants.DEFAULT_ENERGY_THRESHOLD,
         min_region_size=constants.MIN_REGION_SIZE,
         max_region_size=constants.MAX_REGION_SIZE,
-        ext_regions=None,
+        max_continuous_silence=constants.MAX_CONTINUOUS_SILENCE,
+        mode=auditok.StreamTokenizer.STRICT_MIN_LENGTH
+):
+    """
+    Given an input audio/video file, generate proper speech regions.
+    """
+    audio_wav = ffmpeg_utils.source_to_audio(source_file)
+
+    asource = auditok.ADSFactory.ads(
+        filename=audio_wav, record=True)
+    validator = auditok.AudioEnergyValidator(
+        sample_width=asource.get_sample_width(),
+        energy_threshold=energy_threshold)
+    asource.open()
+    tokenizer = auditok.StreamTokenizer(
+        validator=validator,
+        min_length=int(min_region_size * 100),
+        max_length=int(max_region_size * 100),
+        max_continuous_silence=int(max_continuous_silence * 100),
+        mode=mode)
+
+    # auditok.StreamTokenizer.DROP_TRAILING_SILENCE
+    tokens = tokenizer.tokenize(asource)
+    regions = []
+    for token in tokens:
+        # get start and end times
+        regions.append((token[1] * 10, token[2] * 10))
+    asource.close()
+    os.remove(audio_wav)
+    # reference
+    # auditok.readthedocs.io/en/latest/apitutorial.html#examples-using-real-audio-data
+    return regions
+
+
+def sub_gen_speech_regions(
+        source_file,
+        sub_file,
         ext_max_size_ms=constants.MAX_EXT_REGION_SIZE * 1000
-    ):
+):
+    """
+    Given an input audio/video file and subtitles file, generate proper speech regions.
+    """
+    regions = []
+    audio_wav = ffmpeg_utils.source_to_audio(source_file)
+    reader = wave.open(audio_wav)
+    audio_file_length = int(float(reader.getnframes()) / float(reader.getframerate())) * 1000
+    reader.close()
+
+    ext_regions = pysubs2.SSAFile.load(sub_file)
+
+    for event in ext_regions.events:
+        if not event.is_comment:
+            # not a comment region
+            if event.duration <= ext_max_size_ms:
+                regions.append((event.start,
+                                event.start + event.duration))
+            else:
+                # split too long regions
+                elapsed_time = event.duration
+                start_time = event.start
+                if elapsed_time > audio_file_length:
+                    # even longer than the source file length
+                    elapsed_time = audio_file_length
+                while elapsed_time > ext_max_size_ms:
+                    # longer than the max size limit
+                    regions.append((start_time,
+                                    start_time + ext_max_size_ms))
+                    elapsed_time = elapsed_time - ext_max_size_ms
+                    start_time = start_time + ext_max_size_ms
+                regions.append((start_time,
+                                start_time + elapsed_time))
+
+    os.remove(audio_wav)
+    return regions
+
+
+def api_gen_text(  # pylint: disable=too-many-locals,too-many-arguments,too-many-branches,too-many-statements
+        source_file,
+        api_url,
+        regions,
+        api_key,
+        concurrency=constants.DEFAULT_CONCURRENCY,
+        src_language=constants.DEFAULT_SRC_LANGUAGE,
+        dst_language=constants.DEFAULT_DST_LANGUAGE
+):
     """
     Given an input audio/video file, generate subtitles in the specified language and format.
     """
-    audio_wav = source_to_audio(source_file)
-
-    if not ext_regions:
-        asource = auditok.ADSFactory.ads(
-            filename=audio_wav, record=True)
-        validator = auditok.AudioEnergyValidator(
-            sample_width=asource.get_sample_width(),
-            energy_threshold=65)
-        asource.open()
-        tokenizer = auditok.StreamTokenizer(
-            validator=validator,
-            min_length=int(min_region_size * 100),
-            max_length=int(max_region_size * 100),
-            max_continuous_silence=30,
-            mode=auditok.StreamTokenizer.STRICT_MIN_LENGTH
-            | auditok.StreamTokenizer.DROP_TRAILING_SILENCE)
-        tokens = tokenizer.tokenize(asource)
-        regions = []
-        for token in tokens:
-            regions.append((token[1] * 10, token[2] * 10))
-        asource.close()
-    else:
-        regions = []
-
-        reader = wave.open(audio_wav)
-        audio_file_length = int(float(reader.getnframes()) / float(reader.getframerate())) * 1000
-        reader.close()
-
-        for event in ext_regions.events:
-            if not event.is_comment:
-                # not a comment region
-                if event.duration <= ext_max_size_ms:
-                    regions.append((event.start,
-                                    event.start + event.duration))
-                else:
-                    # split too long regions
-                    elapsed_time = event.duration
-                    start_time = event.start
-                    if elapsed_time > audio_file_length:
-                        elapsed_time = audio_file_length
-                    while elapsed_time > ext_max_size_ms:
-                        regions.append((start_time,
-                                        start_time + ext_max_size_ms))
-                        elapsed_time = elapsed_time - ext_max_size_ms
-                        start_time = start_time + ext_max_size_ms
-                    regions.append((start_time,
-                                    start_time + elapsed_time))
-
-    os.remove(audio_wav)
-
     audio_rate = 44100
-    audio_flac = source_to_audio(source_file, rate=audio_rate, file_ext='.flac')
+    audio_flac = ffmpeg_utils.source_to_audio(source_file, rate=audio_rate, file_ext='.flac')
     pool = multiprocessing.Pool(concurrency)
-    converter = SplitIntoFLACPiece(source_path=audio_flac)
-    recognizer = GoogleSpeechToTextV2(language=src_language, rate=audio_rate,
-                                      api_url=api_url_scheme + constants.GOOGLE_SPEECH_V2_API_URL,
-                                      api_key=constants.GOOGLE_SPEECH_V2_API_KEY)
+    converter = ffmpeg_utils.SplitIntoFLACPiece(source_path=audio_flac)
+
+    recognizer = speech_trans_api.GoogleSpeechToTextV2(
+        language=src_language,
+        rate=audio_rate,
+        api_url=api_url,
+        api_key=constants.GOOGLE_SPEECH_V2_API_KEY)
 
     transcripts = []
     if regions:
-        widgets = ["Converting speech regions to FLAC files: ", Percentage(), ' ', Bar(), ' ',
-                   ETA()]
-        pbar = ProgressBar(widgets=widgets, maxval=len(regions)).start()
+        widgets = ["Converting speech regions to FLAC files: ",
+                   progressbar.Percentage(), ' ',
+                   progressbar.Bar(), ' ',
+                   progressbar.ETA()]
+        pbar = progressbar.ProgressBar(widgets=widgets, maxval=len(regions)).start()
         try:
             extracted_regions = []
             for i, extracted_region in enumerate(pool.imap(converter, regions)):
@@ -291,8 +143,11 @@ def generate_subtitles( # pylint: disable=too-many-locals,too-many-arguments,too
                 pbar.update(i)
             pbar.finish()
 
-            widgets = ["Performing speech recognition: ", Percentage(), ' ', Bar(), ' ', ETA()]
-            pbar = ProgressBar(widgets=widgets, maxval=len(regions)).start()
+            widgets = ["Performing speech recognition: ",
+                       progressbar.Percentage(), ' ',
+                       progressbar.Bar(), ' ',
+                       progressbar.ETA()]
+            pbar = progressbar.ProgressBar(widgets=widgets, maxval=len(regions)).start()
 
             for i, transcript in enumerate(pool.imap(recognizer, extracted_regions)):
                 transcripts.append(transcript)
@@ -302,12 +157,16 @@ def generate_subtitles( # pylint: disable=too-many-locals,too-many-arguments,too
             if src_language.split("-")[0] != dst_language.split("-")[0]:
                 if api_key:
                     google_translate_api_key = api_key
-                    translator = GoogleTranslatorV2(dst_language, google_translate_api_key,
-                                                    dst=dst_language,
-                                                    src=src_language)
+                    translator = \
+                        speech_trans_api.GoogleTranslatorV2(dst_language,
+                                                            google_translate_api_key,
+                                                            dst=dst_language,
+                                                            src=src_language)
                     prompt = "Translating from {0} to {1}: ".format(src_language, dst_language)
-                    widgets = [prompt, Percentage(), ' ', Bar(), ' ', ETA()]
-                    pbar = ProgressBar(widgets=widgets, maxval=len(regions)).start()
+                    widgets = [prompt, progressbar.Percentage(), ' ',
+                               progressbar.Bar(), ' ',
+                               progressbar.ETA()]
+                    pbar = progressbar.ProgressBar(widgets=widgets, maxval=len(regions)).start()
                     translated_transcripts = []
                     for i, transcript in enumerate(pool.imap(translator, transcripts)):
                         translated_transcripts.append(transcript)
@@ -331,10 +190,19 @@ def generate_subtitles( # pylint: disable=too-many-locals,too-many-arguments,too
     os.remove(audio_flac)
     timed_subtitles = [(r, t) for r, t in zip(regions, transcripts) if t]
 
-    try:
-        input_m = raw_input
-    except NameError:
-        input_m = input
+    return timed_subtitles
+
+
+def list_to_sub_file(
+        timed_subtitles,
+        output,
+        fps=30.0,
+        subtitles_file_format=constants.DEFAULT_SUBTITLES_FORMAT,
+        input_m=input
+):
+    """
+    Given an input timedsub list, format it and write it to file.
+    """
 
     if subtitles_file_format == 'srt':
         formatted_subtitles = formatters.pysubs2_formatter(
@@ -355,47 +223,13 @@ def generate_subtitles( # pylint: disable=too-many-locals,too-many-arguments,too
             subtitles=timed_subtitles)
     elif subtitles_file_format == 'sub':
         subtitles_file_format = 'microdvd'
-        # ref https://pysubs2.readthedocs.io/en/latest
-        # /api-reference.html#supported-input-output-formats
-        try:
-            command = ["ffprobe", "-v", "0", "-of", "csv=p=0",
-                       "-select_streams", "v:0", "-show_entries",
-                       "stream=r_frame_rate", source_file]
-            use_shell = True if os.name == "nt" else False
-            fps_str = subprocess.check_output(command, stdin=open(os.devnull), shell=use_shell)
-            num_list = map(int, re.findall(r'\d+', fps_str))
-            if len(num_list) == 2:
-                fps = float(num_list[0]) / float(num_list[1])
-            else:
-                print("Can't use ffprobe(ffmpeg) to get video fps.\n"
-                      "Video fps is necessary when output is \".sub\".")
-                fps_str = input_m("Input your video fps. "
-                                  "Any illegal value will regard as \".srt\" instead.")
-                num_list = map(float, re.findall(r'\d+', fps_str))
-                if num_list and num_list[0] > 0.0:
-                    fps = num_list[0]
-                else:
-                    print("Using \".srt\" instead.")
-                    subtitles_file_format = 'srt'
-                    fps = 0.0
-
-        except subprocess.CalledProcessError:
-            print("Can't use ffprobe(ffmpeg) to get video fps.\n"
-                  "It is necessary when output is \".sub\".")
-            fps_str = input_m("Input your video fps. "
-                              "Any illegal value will regard as \".srt\" instead.")
-            num_list = map(float, re.findall(r'\d+', fps_str))
-            if num_list and num_list[0] > 0.0:
-                fps = num_list[0]
-            else:
-                print("Using \".srt\" instead.")
-                subtitles_file_format = 'srt'
-                fps = 0.0
-
         formatted_subtitles = formatters.pysubs2_formatter(
             subtitles=timed_subtitles,
             sub_format=subtitles_file_format,
             fps=fps)
+        # sub format need fps
+        # ref https://pysubs2.readthedocs.io/en/latest
+        # /api-reference.html#supported-input-output-formats
         subtitles_file_format = 'sub'
 
     elif subtitles_file_format == 'mpl2':
@@ -418,12 +252,6 @@ def generate_subtitles( # pylint: disable=too-many-locals,too-many-arguments,too
 
     dest = output
 
-    if not dest:
-        base = os.path.splitext(source_file)[0]
-        dest = "{base}.{langcode}.{extension}".format(base=base,
-                                                      langcode=dst_language,
-                                                      extension=subtitles_file_format)
-
     while os.path.isfile(dest):
         print("There is already a file with the same name"
               " in this location: \"{dest_name}\".".format(dest_name=dest))
@@ -438,12 +266,30 @@ def generate_subtitles( # pylint: disable=too-many-locals,too-many-arguments,too
     return dest
 
 
-def validate(args):
+def validate(args):  # pylint: disable=too-many-branches,too-many-return-statements
     """
     Check that the CLI arguments passed to autosub are valid.
     """
-    if not args.source_path:
-        print("Error: You need to specify a source path.")
+    if args.list_formats:
+        print("List of formats:")
+        for subtitles_format, format_description in sorted(constants.FORMATTERS.items()):
+            print("{sf}\t{fd}".format(sf=subtitles_format, fd=format_description))
+        return False
+
+    if args.list_speech_to_text_codes:
+        print("List of all source language codes:")
+        for code, language in sorted(constants.SPEECH_TO_TEXT_LANGUAGE_CODES.items()):
+            print("{code}\t{language}".format(code=code, language=language))
+        return False
+
+    if args.list_translation_codes:
+        print("List of all destination language codes:")
+        for code, language in sorted(constants.TRANSLATION_LANGUAGE_CODES.items()):
+            print("{code}\t{language}".format(code=code, language=language))
+        return False
+
+    if not args.source_path or not os.path.isfile(args.source_path):
+        print("Error: You need to specify a valid source path.")
         return False
 
     if args.format not in constants.FORMATTERS.keys():
@@ -499,15 +345,21 @@ def validate(args):
         )
         args.max_region_size = constants.MAX_EXT_REGION_SIZE
 
+    if args.max_continuous_silence < 0:
+        print(
+            "Your maximum continuous silence {mxcs} is smaller than 0.\n"
+            "Now reset to {dmxcs}".format(mxcs=args.max_continuous_silence,
+                                          dmxcs=constants.MAX_CONTINUOUS_SILENCE)
+        )
+        args.max_continuous_silence = constants.MAX_CONTINUOUS_SILENCE
+
     return True
 
 
-def main():  # pylint: disable=too-many-branches
+def get_cmd_args():
     """
-    Run autosub as a command-line program.
+    Get command-line arguments.
     """
-    parser = argparse.ArgumentParser()
-
     parser = argparse.ArgumentParser(
         prog=metadata.NAME,
         usage='\n  %(prog)s [options] <source_path>',
@@ -577,12 +429,23 @@ def main():  # pylint: disable=too-many-branches
     )
 
     ogroup.add_argument(
+        '-et', '--energy-threshold',
+        metavar='energy',
+        type=int,
+        default=constants.DEFAULT_ENERGY_THRESHOLD,
+        help="The energy level which determines the region to be detected. "
+             "Ref: https://auditok.readthedocs.io/en/latest/apitutorial.html"
+             "#examples-using-real-audio-data(default: %(default)s)."
+    )
+
+    ogroup.add_argument(
         '-mnrs', '--min-region-size',
         metavar='second',
         type=float,
         default=constants.MIN_REGION_SIZE,
         help="Minimum region size "
-             "when not using external speech regions control(default: %(default)s)."
+             "when not using external speech regions control. "
+             "Auditok uses it. Same docs above(default: %(default)s)."
     )
 
     ogroup.add_argument(
@@ -591,7 +454,31 @@ def main():  # pylint: disable=too-many-branches
         type=float,
         default=constants.MAX_REGION_SIZE,
         help="Maximum region size "
-             "when not using external speech regions control(default: %(default)s)."
+             "when not using external speech regions control. "
+             "Auditok uses it. Same docs above(default: %(default)s)."
+    )
+
+    ogroup.add_argument(
+        '-mxcs', '--max-continuous-silence',
+        metavar='second',
+        type=float,
+        default=constants.MAX_CONTINUOUS_SILENCE,
+        help="Maximum length of a tolerated silence within a valid audio activity. "
+             "Auditok uses it. Same docs above(default: %(default)s)."
+    )
+
+    ogroup.add_argument(
+        '-sml', '--strict-min-length',
+        action='store_true',
+        help="Ref: https://auditok.readthedocs.io/en/latest/core.html#class-summary "
+             "Auditok uses it."
+    )
+
+    ogroup.add_argument(
+        '-dts', '--drop-trailing-silence',
+        action='store_true',
+        help="Ref: https://auditok.readthedocs.io/en/latest/core.html#class-summary "
+             "Auditok uses it."
     )
 
     ogroup.add_argument(
@@ -644,63 +531,91 @@ def main():  # pylint: disable=too-many-branches
         help="Show %(prog)s version and exit."
     )
 
-    args = parser.parse_args()
+    return parser.parse_args()
 
-    if args.list_formats:
-        print("List of formats:")
-        for subtitles_format, format_description in sorted(constants.FORMATTERS.items()):
-            print("{sf}\t{fd}".format(sf=subtitles_format, fd=format_description))
-        return 0
 
-    if args.list_speech_to_text_codes:
-        print("List of all source language codes:")
-        for code, language in sorted(constants.SPEECH_TO_TEXT_LANGUAGE_CODES.items()):
-            print("{code}\t{language}".format(code=code, language=language))
-        return 0
+def main():  # pylint: disable=too-many-branches
+    """
+    Run autosub as a command-line program.
+    """
 
-    if args.list_translation_codes:
-        print("List of all destination language codes:")
-        for code, language in sorted(constants.TRANSLATION_LANGUAGE_CODES.items()):
-            print("{code}\t{language}".format(code=code, language=language))
-        return 0
+    try:
+        input_m = raw_input
+    except NameError:
+        input_m = input
+
+    args = get_cmd_args()
 
     if validate(args):
         if args.http_speech_to_text_api:
-            api_url_scheme = "http://"
+            api_url = "http://" + constants.GOOGLE_SPEECH_V2_API_URL
         else:
-            api_url_scheme = constants.DEFAULT_API_URL_SCHEME
-
+            api_url = "https://" + constants.GOOGLE_SPEECH_V2_API_URL
         try:
+            if args.format == 'sub':
+                fps = ffmpeg_utils.ffprobe_get_fps(
+                    args.source_path,
+                    input_m=input_m)
+                if fps == 0.0:
+                    args.format = 'srt'
+            else:
+                fps = 0.0
+
+            if not args.output:
+                base = os.path.splitext(args.source_path)[0]
+                args.output = "{base}.{langcode}.{extension}".format(base=base,
+                                                                     langcode=args.dst_language,
+                                                                     extension=args.format)
+            elif os.path.isdir(args.output):
+                base = args.output + '.'.split(os.path.basename(args.source_path))[0]
+                args.output = "{base}.{langcode}.{extension}".format(base=base,
+                                                                     langcode=args.dst_language,
+                                                                     extension=args.format)
+                print("Your output is a directory not a file path. "
+                      "Now file path set to {new}".format(new=args.output))
+
             if args.external_speech_regions:
                 print("Using external speech regions.")
-                ext_regions = pysubs2.SSAFile.load(args.external_speech_regions)
-                subtitles_file_path = generate_subtitles(
+                regions = sub_gen_speech_regions(
                     source_file=args.source_path,
-                    concurrency=args.concurrency,
-                    src_language=args.src_language,
-                    dst_language=args.dst_language,
-                    api_url_scheme=api_url_scheme,
-                    api_key=args.api_key,
-                    subtitles_file_format=args.format,
-                    output=args.output,
-                    min_region_size=args.min_region_size,
-                    max_region_size=args.max_region_size,
-                    ext_regions=ext_regions
+                    sub_file=args.external_speech_regions
                 )
 
             else:
-                subtitles_file_path = generate_subtitles(
+                mode = 0
+                if args.strict_min_length:
+                    mode = auditok.StreamTokenizer.STRICT_MIN_LENGTH
+                    if args.drop_trailing_silence:
+                        mode = mode | auditok.StreamTokenizer.DROP_TRAILING_SILENCE
+                elif args.drop_trailing_silence:
+                    mode = auditok.StreamTokenizer.DROP_TRAILING_SILENCE
+
+                regions = auditok_gen_speech_regions(
                     source_file=args.source_path,
-                    concurrency=args.concurrency,
-                    src_language=args.src_language,
-                    dst_language=args.dst_language,
-                    api_url_scheme=api_url_scheme,
-                    api_key=args.api_key,
-                    subtitles_file_format=args.format,
-                    output=args.output,
-                    min_region_size=args.min_region_size,
-                    max_region_size=args.max_region_size
+                    energy_threshold=args.energy_threshold,
+                    min_region_size=constants.MIN_REGION_SIZE,
+                    max_region_size=constants.MAX_REGION_SIZE,
+                    max_continuous_silence=constants.MAX_CONTINUOUS_SILENCE,
+                    mode=mode
                 )
+
+            timed_subtitles = api_gen_text(
+                source_file=args.source_path,
+                api_url=api_url,
+                regions=regions,
+                api_key=args.api_key,
+                concurrency=args.concurrency,
+                src_language=args.src_language,
+                dst_language=args.dst_language
+            )
+
+            subtitles_file_path = list_to_sub_file(
+                timed_subtitles=timed_subtitles,
+                output=args.output,
+                fps=fps,
+                subtitles_file_format=args.format,
+                input_m=input_m
+            )
             print("\nSubtitles file created at \"{}\"".format(subtitles_file_path))
 
         except KeyboardInterrupt: