+++ autosub/constants.py
@@ -13,9 +13,10 @@ DEFAULT_SUBTITLES_FORMAT = 'srt'
 DEFAULT_CONCURRENCY = 10
 DEFAULT_SRC_LANGUAGE = 'en-US'
 DEFAULT_DST_LANGUAGE = 'en-US'
-DEFAULT_API_URL_SCHEME = 'https://'
+DEFAULT_ENERGY_THRESHOLD = 45
 MAX_REGION_SIZE = 6.0
 MIN_REGION_SIZE = 0.5
+MAX_CONTINUOUS_SILENCE = 0.3
 MAX_EXT_REGION_SIZE = 10
 # Maximum speech to text region length in milliseconds
 # when using external speech region control